some changes

+++ user/shmem_test.c
@@ -1,3 +1,4 @@
+#include "types.h"
 #include "user.h"
 
 int
@@ -9,10 +10,10 @@ main(int argc, char *argv[])
     int parent_pid = getpid();
     if(fork() == 0) {
         printf("1. Child process size before shared mapping: %d\n", sbrk(0));
-        int va= map_shared_pages(parent_pid, getpid(), (int)shmem_text, sizeof(shmem_text));
+        uint64 va = map_shared_pages(parent_pid, getpid(), (uint64)shmem_text, sizeof(shmem_text));
         printf("2. Child process size after shared mapping: %d\n",sbrk(0));
         strcpy(va, "Hello daddy");
-        if(unmap_shared_pages(getpid(), va, sizeof(shmem_text))<0){
+        if (unmap_shared_pages(getpid(), va, sizeof(shmem_text)) < 0) {
             printf("Error unmapping shared memory\n");
             exit(1);
         }
